Changed classnames that affecting css, and renamed component Footer to AppFooter
