+++ src/app/App.js
@@ -11,7 +11,7 @@ import Presentation from '../presentation/Presentation';
 import AppHeader from '../common/AppHeader';
 import NotFound from '../common/NotFound';
 import LoadingIndicator from '../common/LoadingIndicator';
-import AppFooter from '../common/Footer';
+import AppFooter from '../common/AppFooter';
 
 import { Layout, notification } from 'antd';
 const { Content, Footer } = Layout;
@@ -84,8 +84,7 @@ class App extends Component {
             </div>
           </Content>
           <Footer className="footer-style">
-              <AppFooter>
-              </AppFooter>
+              <AppFooter />
           </Footer>
         </Layout>
     );
