+++ src/common/AppFooter.css
@@ -1,0 +1,7 @@
+.socials {
+    list-style-type: none;
+  }
+
+.socials a {
+    color: rgb(61, 61, 61);
+}
