--- src/common/Footer.css
@@ -1,7 +1,0 @@
-ul {
-    list-style-type: none;
-  }
-
-a {
-    color: black;
-}
